Add ServiceNow HTTP API client
We add a simple client for communication with the ServiceNow HTTP
API. The client supports basic and OAuth authentication. It wraps
HTTP responses in objects of the convenience `Response` class, which
the caller can use to inspect response status, headers and data.

The client offers methods to support Basic CRUD operations on
ServiceNow resources, i.e. get, post, put and delete.