Add development environment setup instructions
With added instructions, getting up and running should be considerably
easier. We also added Makefile to the repo for convenience.

The final goal is to use make commands inside the CI/CD, but for now,
we did not update the CI/CD pipeline yet because we need some code in
the repo before all our make targets will be functional.