Support for query parameters in the HTTP client's get method
So far, if the caller provided a path argument that included a manually
constructed query string to the HTTP client's get method, things
didn't work as expected due to faulty escaping of the URL.
We fix this by encoding the path part and the query string separately.
We also make passing the query more convenient for the caller - query
params can now be passed as a dictionary, via the query keyword argument.