Add unit testing documentation and fixtures
The testing document provides a basic template of a test file in the
hopes that this will make it a bit easier to write tests. The main
purpose of the documentation and pytest fixtures is to allow content
authors to focus on what to test instead of messing around with
mocks and patches.