Implements ServiceNow incident ticket module
In the current implementation, basic functionalities
creating and deleting incidents are covered with other
incident state options. We also present often used fields
that are important for incidents to be created or in any
other way manipulated. For all other less used or optional
parameters, they can be strung together under other option
in an incident ticket as they are.

In the future, these optional parameters can be brought up
to often used fields if there will be a need.