Construct HTTP client in main method
For ease of testing, we decided to move HTTP client construction
from `run` to `main` method. From now on, every module's `run`
method should accept the client instance as the second argument.
This allows us to get rid of mocking the client from within unit
tests of the `run` method. Instead, we now have a simple fixture that
returns the mocked client instance, which we can inject to the run
method.

We also update the relevant developer documentation affected by
this change.