Add query parameter all HTTP methods
All ServiceNow table API methods take query parameters, but we only
exposed this functionality for the GET method. This commit fixes this
omission and adds query parameter to all wrapped HTTP methods.

The main reason for this change is the sysparm_exclude_reference_link
query parameter that allows us to suppress link listing in API
responses, keeping the response flat and nicer to use in further
processing.