Encapsulate table querying into a class
The main reason for this is convenience. Since most of the modules
will not touch the raw HTTP client, we might as well feed the more
abstract client in the first place.

For easier testing, we also provided a pytest fixture that can be used
when unit testing the module logic.