Helper for ensuring the presence of required fields
Before state of a record can be modified, ServiceNow API may require
that a record has certain fields set. In order to make things more
flexible and not simply force the user to provide a set of mandatory
parameters with every task that involves state transition, we add a
helper that ensures specified parameters are given either in
the module parameters, or already set on the remote record.