Add initial inventory plugin
The inventory plugin adds hosts from records of the specified
ServiceNow table. The users can customize which table
to query, which columns to retrieve as hostvars, and what columns
should be used for ansible_host and inventory hostname.

Inventory groups can be generated in one of two ways:
a) The user specifies the groups explicitly, or
b) Groups are created automatically, based on the values of
   the provided columns.

To limit the records from which Ansible hosts will be created,
we support simple filtering of column values with 'includes'
and 'excludes'. We support specifying criteria in this manner
for both user-specified groups as well as for the automatically
generated ones.