Implements ServiceNow change request module
In the current implementation, basic functionalities
creating and deleting change requests are covered with
other state and type options. We also present often used
fields that are important for change requests to be created
or in any other way manipulated. For all other less used
or optional parameters, they can be strung together under
other field in change request as they are.

In the future, these optional parameters can be brought up
to often used fields if there will be a need.