Reuse validation helper in incident module
Now that we have a dedicated helper for checking the presence of
required values, we might as well use it in all of our modules.

Changing the implementation to use the newly-introduced helper was
straightforward, but we did discover a subtle bug in our incident
module while changing the validation code. If the task was set to
create a new incident in resolved or closed state without providing
required data, we failed to detect this omission.

We fixed the aforementioned issue and added an integration test to
make sure it stays fixed.