Add configuration item module
We add configuration_item module that manages records in the
ServiceNow configuration item tables.

To decide the target ServiceNow CMDB table to insert new records
into, the module consults the value of `sys_class_name` module
parameter, or uses the default value `cmdb_ci`.
For the existing records, the module first queries the parent
`cmdb_ci` table in order to obtain the configuration item's
`sys_class_name`, which points to the extended ServiceNow CMDB
table. It then modifies the record in that table.