Rename group inventory parameter
Because some built-in inventory helpers already use the group
parameter, we renamed our group parameter to named_groups.

We could also remove the group parameter altogether and use the
constructed inventory to simulate it, but we left it in because the
named groups implementation can be more efficient when we want to
retrieve only a specific subset of items (the resource filtering is
done on the backend).