Refactor add_host method
The add_host method was accessing the global configuration and that
coupled it tightly to the underlying inventory implementation.

Changes in this commit moved the global data access into parameters that
the caller needs to provide, making it possible to properly test the
functionality.

While we were at it, we also made sure the inventory construction fails
if user misconfigured the Ansible host source. Do note that empty value
is still OK as ansible_host value (inventory can be used in a wide
variety of ways and quite a lot of them do not need to have ansible_host
set).