Deprecate duplicated inventory options
With the introduction of the constructed features to the inventory
plugin, a lot of existing functionality is not needed anymore since it
can be replaced by the constructed features.

This commit marks options that control the obsolete functionality as
deprecated. Inventory plugin will also print a warning if it detects
deprecated features are being used.