Deprecate duplicated inventory options (#97)
Deprecate duplicated inventory options

With the introduction of the constructed features to the inventory plugin, a lot of existing functionality is not needed anymore since it can be replaced by the constructed features.
This commit marks options that control the obsolete functionality as deprecated. The inventory plugin will also print a warning if it detects deprecated features being used since Ansible has some issues printing deprecation notices from plugins that are lazy-loaded.

Reviewed-by: Matej Pevec <None>
Reviewed-by: Tadej Borovšak <None>
Reviewed-by: Jill R <None>
Reviewed-by: Alina Buzachis <None>
Reviewed-by: None <None>