Add modules for managing problem tasks
In the current implementation, basic functionalities creating and
deleting problem tasks are covered.  We also present often used fields
that are important for problem tasks to be created or in any other way
manipulated. For all other less used or optional parameters, they can
be strung together under other field in the problem tasks as they are.

There is an issue with the state transition from Assess to Work In
Progress and to Closed.  When manually transitioning to Work In
Progress state fields started_at and stared_by are field in but the
module can not change to that state even though these two parameters
are filled.