Remove assumed /api/now base path from the HTTP client (#131)
Remove assumed /api/now base path from the HTTP client

SUMMARY
ServiceNow Update Sets allow for extending the functionality of ServiceNow by, for example, obtaining and formatting results from joined tables. Through the ServiceNow REST API the records from Update Sets can be programmatically obtained. Currently, in the HTTP client it is assumed that all API endpoints start with base path /api/now, which might not be true in case of Updated Sets or in general when the requests might refer to the records outside namespace now.
This update removes the assumption about the base URL path from the HTTP client, by accepting any path, which makes it more generic. While it does not directly address any issue, it shall be seen as a prerequisite to issue #108, the resolution of which can be done by using a particular Updated Set, which is accessible on URL path starting with /api/snc.
ISSUE TYPE

Feature Pull Request (a prerequisite to)

COMPONENT NAME
This update does not introduce any new components.
ADDITIONAL INFORMATION
The primary target of this update is request method of class Client, located in file plugins/module_utils/client.py, where the base path is changed to "/". Consequentially, all the code relying upon the request method, either directly or indirectly, is affected. Accordingly, the TableClient and AttachmentClient have been updated. The unit tests for Client, TableClient and AttachmentClient have also been adapted to the change in the base URL path.

Reviewed-by: Tadej Borovšak <None>
Reviewed-by: Abhijeet Kasurde <None>
Reviewed-by: None <None>