Bugfix - mapping may be None (#138)
Bugfix - mapping may be None

Mapping is defined in module definition and is optional. In case that it
is not listed in playbook it has default value None, which was not covered.
Case that it is not in dictionary or partial was allready covered.
SUMMARY


ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

Mapping
ADDITIONAL INFORMATION

Reviewed-by: Uroš Paščinski <None>
Reviewed-by: Abhijeet Kasurde <None>
Reviewed-by: None <None>