Return nothing on empty string in add_host
ServiceNow API doesn't like it when sysparm_limit is set to a high
value like 1,000,000 and the amount of records needing to be
returned is 90,000. What I've found is that an empty string gets
inserted into either at the end or somewhere with the list of
records.

My reasoning for this commit is that if its an empty string or if
the record is of NoneType, then the code should just skip it since
it is a bad data instead of failing outright.